Поправки для Result
